Ensure open() works even if tiger.pomdp does not exist
I added the "w" argument, because in a fresh installation (depending on your path) the "tiger.pomdp" file will not exist. At least it didn't when I tried this in my setup so adding the "w" argument would prevent Julia newbies to try to figure out why the open("tiger.pomdp") does not work.